// SDKs/twilio.com/twilio_voice_v1/1.36.0/typescript/src/sdk/models/shared/voicev1dialingpermissionsdialingpermissionscountry.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class VoiceV1DialingPermissionsDialingPermissionsCountry extends SpeakeasyBase {
  @Metadata({ data: "json, name=continent" })
  continent?: string;

  @Metadata({ data: "json, name=country_codes" })
  countryCodes?: string[];

  @Metadata({ data: "json, name=high_risk_special_numbers_enabled" })
  highRiskSpecialNumbersEnabled?: boolean;

  @Metadata({ data: "json, name=high_risk_tollfraud_numbers_enabled" })
  highRiskTollfraudNumbersEnabled?: boolean;

  @Metadata({ data: "json, name=iso_code" })
  isoCode?: string;

  @Metadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @Metadata({ data: "json, name=low_risk_numbers_enabled" })
  lowRiskNumbersEnabled?: boolean;

  @Metadata({ data: "json, name=name" })
  name?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}